pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "grok-build-bridge"
version = "0.1.0"
description = "Turn any Grok-generated codebase into a safely deployed X agent via one YAML file."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
authors = [{ name = "Jan Solo", email = "jan@agentmind.cloud" }]
maintainers = [{ name = "AgentMindCloud" }]
keywords = ["grok", "xai", "x", "twitter", "agent", "bridge", "deployment"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Build Tools",
]
dependencies = [
    "xai-sdk>=1.0",
    "pyyaml>=6",
    "rich>=13",
    "typer>=0.12",
    "httpx>=0.27",
    "jsonschema>=4.21",
    "tenacity>=8.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio",
    "pytest-cov>=5",
    "mypy>=1.10",
    "ruff>=0.6",
    "build>=1.2",
    "types-PyYAML",
    "types-jsonschema",
]

[project.urls]
Homepage = "https://github.com/AgentMindCloud/grok-build-bridge"
Repository = "https://github.com/AgentMindCloud/grok-build-bridge"
Issues = "https://github.com/AgentMindCloud/grok-build-bridge/issues"
Ecosystem = "https://github.com/AgentMindCloud/grok-install-ecosystem"

[project.scripts]
grok-build-bridge = "grok_build_bridge.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["grok_build_bridge"]
# ``templates/`` and ``schema/`` live under the package directory, so
# hatchling picks them up via ``packages`` above. No ``force-include``
# needed — adding one would create duplicate entries in the wheel.

[tool.hatch.build.targets.sdist]
include = [
    "grok_build_bridge",
    "tests",
    "docs",
    "README.md",
    "LICENSE",
    "CHANGELOG.md",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "W"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"grok_build_bridge/cli.py" = ["B008"]
"grok_build_bridge/_banner.py" = ["E501", "RUF001"]

[tool.ruff.format]
line-ending = "lf"
quote-style = "double"

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true
# Typer stores runtime metadata in ``Option`` / ``Argument`` defaults that
# mypy cannot type-check; allow those boundaries rather than littering the
# CLI with per-line ignores.
[[tool.mypy.overrides]]
module = "grok_build_bridge.cli"
disallow_untyped_decorators = false
disallow_any_decorated = false

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
]

[tool.coverage.run]
source = ["grok_build_bridge"]
branch = true
omit = [
    "grok_build_bridge/_banner.py",
]

[tool.coverage.report]
skip_empty = true
show_missing = true
precision = 1
exclude_also = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "\\.\\.\\.",
]
